requirements.txt
discord.py
flask
aiosqlite
aiohttp
schedule